migrate: add session approval service

+++ crates/agent-core-v2/src/agent/user_tool/service.rs
@@ -61,7 +61,7 @@ impl ExecutableTool for UserToolExecutable {
                     }),
                     origin: Some(InteractionOrigin {
                         agent_id: None,
-                        turn_id: Some(context.turn_id),
+                        turn_id: Some(context.turn_id as f64),
                     }),
                 };
                 match abortable(interaction.request(request), &context.signal).await {
