+++ crates/agent-core-v2/src/session/interaction/service.rs
@@ -127,7 +127,7 @@ impl SessionInteractionService {
     }
 
     // Original: cancelPendingForTurn().
-    pub async fn cancel_pending_for_turn(&self, turn_id: u64) {
+    pub async fn cancel_pending_for_turn(&self, turn_id: f64) {
         let cancelled = {
             let mut state = self.state.lock().await;
             let ids = state
@@ -252,7 +252,7 @@ mod tests {
                         payload: serde_json::Value::Null,
                         origin: Some(InteractionOrigin {
                             agent_id: None,
-                            turn_id: Some(3),
+                            turn_id: Some(3.0),
                         }),
                     })
                     .await
@@ -260,7 +260,7 @@ mod tests {
         };
         tokio::task::yield_now().await;
         assert_eq!(service.list_pending(None).await.len(), 1);
-        service.cancel_pending_for_turn(3).await;
+        service.cancel_pending_for_turn(3.0).await;
         assert_eq!(
             wait.await.unwrap(),
             serde_json::json!({"cancelled": true, "reason": "turn_ended"})
